Use a little dd action to wipe out any LVM info that may be lurking on the SD card.
When copying the 900MiB image on an SD card that has already had ev3dev installed,
the LVM information for the 3rd partition is still on the SD card and gets picked
up by LVM, leading to duplicate volume group errors. This causes the firstboot
script to fail, which causes even more problems.

It looks like the LVM information is in the first 12KiB of the /dev/mmcblk0p3
partition, so that is all we need to wipe out (by writing zeros).

Thanks to @G33kDude for providing a broken SD card image to test.

Issue ev3dev/ev3dev#267