Provide temporary awk during package config.
Some packages require awk to be in the path during configuration,
however awk (gawk, mawk, ...) has not been configured yet.

This solution provides a temporary /usr/local/bin/awk symlinked
by to /usr/bin/awk at the beginning of configuration, and removes
it at the end of configuration.

The advantage of this solution over an awk script are:
  * do not have to maintain scripts in multiple board directories.
  * avoid potential issues with expansion of quoted arguments.
  * avoid other interpreters to handle the script (e.g bash, perl, ...)
  * Debian will eventually install a symlink, so we are doing the same.