Update namespace prefixes. See also issue #20: https://github.com/ev3dev/brickstrap/issues/20
Most of these functions are for the internal use by brickstrap and should not be used directly by user code (e.g. hooks) so now they are marked as part of the private brickstrap namespace.

Also add some initial work on documentation. This is mainly to ensure work on documenting this properly is 'set in motion' ... help very much welcome!