Use brickstrap.conf.d to store results of merging components
This changes things around a bit so that on the first call to brickstrap,
it creates a new directory named brickstrap.conf.d and fill is with the
combined files of all of the components specified on the command line.

This allows you to see what the actual source files for creating the image
are. It also means you don't have to rember which project and components
you used on subsequent calls to brickstrap, which is very useful when
you are using it as a build environment that you keep around for a long
time.

Fixes #24