Move check for deleting existing conf dir on new command
The existing conf was being read in brp_init_env when running a new
command which would overwrite the component variables. This would
cause brickstrap to recreate the old configuration instead of creating
a new configuration using the command line options.

Moving the check to brp_validate_cli_options deletes the existing
configuration before we get to brp_init_env. Now we get the desired
behavior.