ev3dev-jessie: fix rpi-base/device-tree hook
Apparently there is a *.old directory with overlays in it that gets
picked up by the glob. This causes cp to be unhappy a and fail with
"will not overwrite just-created" because it is trying to copy two
of each file to the same place.

Fixed up the glob to exclude the *.old directory